backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["surister <surister98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
